requirements.txt
gradio
google-generativeai
pymupdf
matplotlib
numpy
